Grid region default region position
- Provide configurable option to set initial region location and webui map center

NOTE: Upstream provider WhiteCore-Sim has done this, however it is unclear why they needed to do so.  Because it is unclear why this is needed but does appear to be logical with the way the Second Galaxy Development Team's comprehensive goals for our architectures will work in the future and it would in fact be somewhat logical for SpaceGate I am committing this to a branch for review by developers.

Signed-Off-By: BruceStarfinder<bruce@secondgalaxy.com>