Initial cleanup of BulletSim
- CLean up unnecessary whitespaces
- Fix licensing as necessary
- Remove commented out code that is either no longer used or depreciated
- CLean up trailing whitespaces
- General code cleanup
- Update the TODO.txt file

NOTE: While it is reasonable to keep commented out code that is depreciated for historical values, it actually contributes to a bloated code based in the binary files upon being built.  This potentially will slow down the services overall.  This is what repositories are for as it keeps historical notes.

Signed-Off-By: QueenStarfinder<queen@secondgalaxy.com>