Add ReactOS to SupportedOS.md and SupportedOS.txt files
NOTE: ReactOS is not yet tested by the Second Galaxy Development Team for Virtual Universe.  The operating system's developers indicate it is an open source alternative to the Windows Operating Systems based on Windows and not Unix.  Because it is currently in Alpha development stages it does not meet the development team's standards for providing Virtual Universe support to run our architectures on the operating system.

Signed-Off-By: BritanyannStarfinder<britanyann@secondgalaxy.com>