Temporarily restore XMLRPC service to Region
- Restores the XMLRPC service to the region servers temporarily

NOTE: While I agree that the XMLRPC Service should be a grid service, it would appear that the move of this may have caused issues with Issue #24 where water claims to be missing but is not and Issue #25 where the sky is missing.  A more permanent fix is in the pipeline but this should help allievate any issues while we investigate this further.

Signed-Off-By: EmperorStarfinder<emperor@secondgalaxy.com>