Format cleanup (Major Commit)
NOTE: This cleans up the formatting which was at best garbled up.  Some of this was likely caused by updates to the Github service which Github did while some in fact did come from various upstream providers who need to clean up their formatting.  Each file is now in process of having any and all garbage removed.  As these files are completed they will be pushed to the repo

Signed-Off-By: NoahStarfinder<noah@secondgalaxy.com>