Increase ReadBuffer size to 2MB
- This increases the ReadBuffer size to 2MB in the DecompressOsd function in Meshmerizer

NOTE: It should be noted that while I am increasing the buffer size, it actually will not make much difference if any at all when it comes to meshing of objects, terrain, etc.  This would more or less be a change that has no real value or meaning and is not really needed.

Signed-Off-By: QueenStarfinder<queen@secondgalaxy.com>