Implement display of staff UserTypes in profile
- Implement the ability to display the UserTypes Mentor, Staff, Assistant Developer, and Core Developer in the user's profiles in-world and in the builtin webui.

NOTE: Devil Starfinder originally requested this.  His rationale was that Linden Labs does this with their employees on Secondlife.  After review we find this is in fact the case and it does make sense.  The vote on this request at the last Core Developers meeting of the Second Galaxy Development Team was 10-0 in favor.

Signed-Off-By: NoahStarfinder<noah@secondgalaxy.com>