Remove unnecessary UserType
- Removes the Guest UserType as it is not really needed
- Changes the dispay of resident in the user's profile to citizen to match the architecture used by Second Galaxy

NOTE: The UserType Guest is never really used even in Secondlife.  It was intended only to act as a very temporary UserType placeholder until the user logs in for the first time at which point their UserType changes to Resident.  At last check tonight Secondlife no longer does this as verified by creating a new user account on secondlife and checking the user's profile prior to their first login shows Resident as the UserType and not Guest.  Therefore it is not really needed so we remove it and deem it unnecessary.

Signed-Off-By: NoahStarfinder<noah@secondgalaxy.com>