[SECURITY FLAG]: Memory Issues
- Tweak the messaging for memory usage

NOTE: ATTN Windows Users - The .Net Framework 4.6.1 has an issue that causes more memory to be used then is aboslutely necessary.  Microsoft is aware of this and will be rolling out a fix in the January 2017 updates to address this isue.  While this will always trigger the security flags for the Second Galaxy Development Team it is not a security flag you should be worried about.  It does not cause any backdoor issues we are aware of for users.  Linux users are not affected by memory spikes being reported when users log out or view world map or use other services in the architecture.

Signed-Off-By: PirittaStarfinder<piritta@secondgalaxy.com>