Fix MaxThreadCount
- Increase the configured number to be 100 instead of 50 in Config/Sim/Scripting/VirtualScript.ini

NOTE: If you are using Microsoft Windows you may need to keep this at MaxThreadCount = 50 due to how Windows handles threading compared to Linux or Mac OS X.

NOTE 2: To answer Issue https://github.com/Virtual-Universe/Virtual-Dev/issues/33

It is a carry over from Opensim 0.6.3 and was likely changed in Aurora-Sim in the engine code but forgotten in the configuration file.  No Evidence to suggest that it got changed in Upstream-WhiteCore.

Signed-Off-By: BritanyannStarfinder<britanyann@secondgalaxy.com>